Add top level IPv4 module, move InternetChecksum out of IP module
